requirements.txt
cython
matplotlib
pillow
pycocotools
scipy
tensorboardX
torch
torchvision
tqdm
ilabs.curate  # for scarlet dataset
opencv-python
